Added Gameplay and Level Prefabs, New InputSystem
Added prefabs for button, pressure plate, level segment and player as well as scripts that go along with the button, pressure plate and player. Also added the new InputSystem to the project for the player inputs. Also made a quick sample level as a reference for how the prefabs are used/work.