Fixed bug with pressure plate and boulder, minor adjustment in wallScript
Fixed bug with pressure plate going haywire when boulder and player were both on it. Restructured the collision checks that it does. It now keeps a list of objects that collide with it. When that list is empty it knows it is not being pushed down and should deactivate its signalers.