Start/End Point Teleports
Created a new prefab called "Start Point" which is basically just a transform marker

Changed EndPoint script so it takes reference and "connects" to another transform (Start Point), and if player walks onto EndPoint it transfers to the connected StartPoint. EndPoint can also be set to "isExit", in which case the application closes (only works in Unity editor). Each EndPoint also takes a cameraPanDirection which has to be manually set to let the camera know which way to pan.