CameraTrigger Fix
Script now stores last activated trigger and deactivates the collider on it, when another trigger is hit the previous trigger collider gets re-activated and the last hit collider deactivated. Should solve issues with accidentally panning the camera multiple times in one direction if hovering on the trigger line with the player.